Article Version Management

+++ EasyReadsDAL/Repos/ArticleRepo.cs
@@ -28,8 +28,11 @@ namespace EasyReadsDAL.Repos
         public void DeleteArticle(int id)
         {
             var article = _context.Articles.Find(id);
-            _context.Articles.Remove(article);
-            _context.SaveChanges();
+            if (article != null)
+            {
+                _context.Articles.Remove(article);
+                _context.SaveChanges();
+            }
         }
 
         public List<Article> GetAllArticles()
@@ -44,7 +47,11 @@ namespace EasyReadsDAL.Repos
 
         public List<Article> GetAllArticlesOfTopic(int topicId)
         {
-            throw new NotImplementedException();
+            var articles = (from article in _context.Articles
+                            where article.TopicId == topicId && article.Audience.Equals("Public")
+                            select article).ToList();
+
+            return articles;
         }
 
         public Article? GetArticle(int id)
@@ -55,7 +62,11 @@ namespace EasyReadsDAL.Repos
 
         public List<Article> GetFollowersOnlyArticlesOfAuthor(string username)
         {
-            throw new NotImplementedException();
+            var articles = (from article in _context.Articles
+                            where article.WrittenBy.Equals(username) && article.Audience.Equals("Public") || article.Audience.Equals("Followers")
+                            select article).ToList();
+
+            return articles;
         }
 
         public List<Article> GetPublicArticlesofAuthor(string username)
@@ -70,38 +81,76 @@ namespace EasyReadsDAL.Repos
 
         public void IncBookmarkCount(int id)
         {
-            throw new NotImplementedException();
+            var article = _context.Articles.Find(id);
+            if(article != null)
+            {
+                article.BookmarksCount++;
+            }
+            _context.SaveChanges();
         }
 
         public void IncCommentsCount(int id)
         {
-            throw new NotImplementedException();
+            var article = _context.Articles.Find(id);
+            if (article != null)
+            {
+                article.CommentsCount++;
+            }
+            _context.SaveChanges();
         }
 
         public void IncLikesCount(int id)
         {
-            throw new NotImplementedException();
+            var article = _context.Articles.Find(id);
+            if (article != null)
+            {
+                article.LikesCount++;
+            }
+            _context.SaveChanges();
         }
 
         public void DecBookmarkCount(int id)
         {
-            throw new NotImplementedException();
+            var article = _context.Articles.Find(id);
+            if (article != null)
+            {
+                article.BookmarksCount--;
+            }
+            _context.SaveChanges();
         }
 
         public void DecCommentsCount(int id)
         {
-            throw new NotImplementedException();
+            var article = _context.Articles.Find(id);
+            if (article != null)
+            {
+                article.CommentsCount--;
+            }
+            _context.SaveChanges();
         }
 
         public void DecLikesCount(int id)
         {
-            throw new NotImplementedException();
+            var article = _context.Articles.Find(id);
+            if (article != null)
+            {
+                article.LikesCount--;
+            }
+            _context.SaveChanges();
         }
 
 
         public void UpdateArticle(Article article)
         {
-            throw new NotImplementedException();
+            var data = _context.Articles.Find(article.ArticleId);
+            if(data != null)
+            {
+                data.Title = article.Title;
+                data.Content = article.Content;
+                data.NumberOfWords = article.NumberOfWords;
+                data.ModifiedAt = article.ModifiedAt;
+                _context.SaveChanges();
+            }
         }
 
         public List<Article> GetAllPublicArticles()
@@ -121,13 +170,54 @@ namespace EasyReadsDAL.Repos
             return _context.Articles.FromSqlRaw(@"
                     SELECT TOP 10 *
                     FROM Articles
-                    ORDER BY LikesCount DESC, CommentsCount DESC, BookmarkCount DESC
+                    ORDER BY LikesCount DESC, CommentsCount DESC, BookmarksCount DESC
                 ").ToList();
         }
 
         public List<Article> GetTopArticlesByTopic(int topicId)
         {
-             throw new NotImplementedException();
+            return _context.Articles.FromSqlRaw(@"
+                    SELECT TOP 10 *
+                    FROM Articles WHERE opicId = {0}
+                    ORDER BY LikesCount DESC, CommentsCount DESC, BookmarksCount DESC
+                ", topicId).ToList();
+        }
+
+        public void CreateArticleVersion(ArticleVersion version)
+        {
+            _context.ArticleVersions.Add(version);
+            _context.SaveChanges();
+        }
+
+        public void DeleteArticleVersion(int id)
+        {
+            var data = _context.ArticleVersions.Find(id);
+            if (data != null)
+            {
+                _context.ArticleVersions.Remove(data);
+                _context.SaveChanges();
+            }
+        }
+
+        public ArticleVersion? GetArticleVersion(int id)
+        {
+            return _context.ArticleVersions.Find(id);
+        }
+
+        public List<ArticleVersion> GetAllArticleVersions(int articleId)
+        {
+            var data = (from ar in _context.ArticleVersions where ar.ArticleId == articleId select ar).ToList();
+            return data;
+        }
+
+        public void DeleteAllVersions(int articleId)
+        {
+            var data = GetAllArticleVersions(articleId);
+            foreach (var version in data)
+            {
+                _context.ArticleVersions.Remove(version);
+            }
+            _context.SaveChanges();
         }
     }
 }
